Separate each task to individual modules.
